requirements.txt
gradio
sentence-transformers
faiss-cpu
PyMuPDF
openai
numpy
